+++ CHANGELOG.md
@@ -1,6 +1,21 @@
 # Changelog
 
-## [v.2.3.0] - August 1 2020
+## [v2.x.x] - forthcoming in Sept 2020?
+
+### Added
+
+### Changed
+
+- currentdrawing() function to return the current drawing if there is one
+- show() for new Drawing() eg in REPL
+- background() preserves graphics state
+- imagematrix() bugs in alpha
+
+### Removed
+
+### Deprecated
+
+## [v2.3.0] - August 1 2020
 
 ### Added
 
@@ -14,7 +29,7 @@
 
 ### Deprecated
 
-## [v.2.2.1] - patch 14 July 2020
+## [v2.2.1] - patch 14 July 2020
 
 ### Added
 
@@ -28,7 +43,7 @@
 
 ### Deprecated
 
-## [v.2.2.0] - 2 July 2020
+## [v2.2.0] - 2 July 2020
 
 ### Added
 
@@ -43,7 +58,7 @@
 
 ### Deprecated
 
-## [v.2.1.0] - 2020 June 18
+## [v2.1.0] - 2020 June 18
 
 ### Added